+++ releasenotes.md
@@ -13,6 +13,7 @@ The game is built using Pygame and is designed to be easy to play and visually a
 
 #### Basic gameplay
 - **[Issue #1](https://github.com/ChristosHadjigeorghiou1996/PySnakey/issues/1)** - Implement classes and interaction</li>
+- **[Issue #4](https://github.com/ChristosHadjigeorghiou1996/PySnakey/issues/4)** - Implement high score mode</li>
 
 #### Repository structure
 - **[Issue #6](https://github.com/ChristosHadjigeorghiou1996/PySnakey/issues/6)** - Implement releasenotes and milestones</li>